# docs/swagger.yaml
definitions:
  domain.Description:
    properties:
      functions:
        items:
          $ref: '#/definitions/domain.ExperimentFunction'
        type: array
      lang:
        type: string
      name:
        type: string
    type: object
  domain.Design:
    properties:
      description:
        $ref: '#/definitions/domain.Description'
      id:
        type: string
      js_filename:
        type: string
      name:
        type: string
      wasm_filename:
        type: string
    type: object
  domain.ExperimentFunction:
    properties:
      args:
        items:
          type: number
        type: array
      function:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/design:
    post:
      consumes:
      - application/json
      description: Создает новый план эксперимента
      parameters:
      - description: Данные плана эксперимента
        in: body
        name: design
        required: true
        schema:
          $ref: '#/definitions/domain.Design'
      produces:
      - application/json
      responses:
        "201":
          description: Создано
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: создает новый план эксперимента
      tags:
      - Design
  /api/design/{id}:
    delete:
      description: Удаляет план эксперимента по ID.
      parameters:
      - description: ID плана эксперимента
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Удалено
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: удаляет план эксперимента
      tags:
      - Design
  /api/designs:
    get:
      description: Возвращает список планов экспериментов.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Design'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: получает список планов экспериментов
      tags:
      - Design
swagger: "2.0"